Made a js file for local storage
